requirements.txt
appdirs==1.4.4
astroid==2.4.2
beautifulsoup4==4.9.3
black==20.8b1
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
decorator==4.4.2
idna==2.10
isort==5.5.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.0
pycparser==2.20
pylint==2.6.0
python-dotenv==0.14.0
python-Levenshtein-wheels==0.13.1
python-telegram-bot==12.8
redis==3.5.3
regex==2020.9.27
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
tornado==6.0.4
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
wrapt==1.12.1
